pyproject.toml
[project]
name = "financial-report-ai-assistant"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.15"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "streamlit (>=1.48.0,<2.0.0)",
    "uvicorn[standard] (>=0.35.0,<0.36.0)",
    "pymupdf (>=1.26.3,<2.0.0)",
    "pandas (>=2.3.1,<3.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "requests (>=2.32.4,<3.0.0)",
    "langchain (>=1.0.8,<2.0.0)",
    "langchain-openai (>=1.0.3,<2.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "faiss-cpu (>=1.13.0,<2.0.0)",
    "tiktoken (>=0.12.0,<0.13.0)",
    "sentence-transformers (>=5.1.2,<6.0.0)",
    "langchain-huggingface (>=1.0.1,<2.0.0)",
    "langchain-community (>=0.4.1,<0.5.0)",
    "tabulate (>=0.9.0,<0.10.0)",
    "llama-parse (>=0.6.81,<0.7.0)"
]

[tool.poetry]
packages = [{include = "financial_report_ai_assistant", from = "src"}]


[tool.poetry.group.dev.dependencies]
poethepoet = "^0.37.0"
jupyterlab = "^4.4.6"
ipykernel = "^6.30.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poe.tasks]
start = { cmd = "uvicorn financial_report_ai_assistant.api.main:app --reload --app-dir src", help = "Starts the FastAPI development server." }